feat: add scroll-to-top button and refactor app sidebar into shared components
